Introduced another build mode called bench.
Since it turned out that the performance implications of
runtime-controllable logging are measurable and too large to ignore,
but I also want to keep the flexibility and user-friendliness of
runtime-controllable logging, I've decided to solve it with another
build mode.

The idea is to only control logging via the environment variable LOG in
debug, release, and coverage, but hardcode it to Info,Error in bench
mode.

I think this is the best compromise here, because it allows us to avoid
the trade-off between debuggability and performance. That is, without
the bench mode we always have to think about whether another log
statement might reduce the performance too much or whether it's worth
it given the better debuggability provided by the log statement. That's
a very bad trade-off and therefore we should not have that.

Thus, I think the better idea is to always work in debug or release mode
and only for our benchmarks, switch to bench.

This begs the question whether it makes it harder to analyze problems in
bench mode now that we cannot use logging in that mode. I think it is
okay, because if it's a timing problem logging doesn't help anyway. If
it's not a timing problem we should also have the same problem in
release mode, where we can enable logging.