Overhauled the server API.
The previous version led to a lot of boilerplate code and was therefore
pretty hard to use and to explain. For that reason, I've now overhauled
the server API to better support its common usage.

The main idea is to allow the registration of handler functions for both
capability exchanges and messages from clients. At first, this relieves
server developers from doing the same every time: fetch the opcode from
the beginning of the message and then check what function to call.

Second and more importantly, it provides the foundation for a better
structure for capability exchanges. Previously, we typically had a
single obtain/delegate function in every server that did some common
actions at the beginning and end and then had a big match expression in
the middle for the different operations. This had not only the downside
that it was pretty hard to read and write, but also made capability
handler work completely different from message handlers. The new API
uses the same approach for both types: a handler function is registered
for a particular opcode and called whenever this operation is
requested.

An important design goal of the new API is also modularity. It should be
possible to use Server, but implement capability exchanges and message
exchanges with clients completely different. For example, it is
imaginable to have a server that only sends events to clients. In this
case, we don't need any request handling on the server side. To that
end, the foundation of the server API is provided by the Server struct,
which is responsible for handling the interaction with the kernel
(registration, session creation, capability exchanges, etc.). The
Server is customized by a Handler trait, which defines how sessions are
opened and closed and is responsible for capability exchanges over
these sessions. The typically used implementation of Handler is the
RequestHandler, which is the primary part of the API that has been
refactored in this commit. The RequestHandler implements the way we
handle client requests in pretty much all servers we have so far and
supports the registration of capability handlers and message handlers.

There are still a few TODOs open, which will be addressed in the next
commits, but the basics work and simplified the implementation of most
servers significantly. In particular, simple servers are now much
simpler to implement, which makes it much easier to explain the API in
lectures, tutorials, etc.