Improved closing of sessions via client request.
The previous rather hacky version was to have a global variable in every
server that we put the session id to remove in and a check in the
server loop whether there is a session to remove.

I've now extended the RequestSession trait to contain a method "alive"
that can be overridden by servers. This gives them the chance to set
sessions to dead on behalf of a client request, which causes the
request handler to remove the session after this request.