boot: increased number of m3fs sessions.
In contrast to earlier, we are now also initializing the session
container with the given maximum number of sessions. At least in m3fs,
we previously only accounted for the number of meta sessions and always
allocated a session container with the max. capacity.

I think we should also limit the session container accordingly and
account for that in the boot scripts.