Let RequestHandler create the ServerSession objects.
This simplifies the per-server code significantly and also allows the
RequestHandler to control how they created. For example, it wants to
make all sessions that are created from existing ones (not via the
resmng/kernel) auto-close, whereas this is not required by sessions
created via the resmng/kernel.

This change also removes the need to keep the server selector around
somewhere (for which we used a StaticCell in various servers).