Replaced usage of int_enum with num_enum for opcodes.
This makes the registration of message/capability handlers nicer,
because we can simply specify the operation instead of the operation
with a ".val" suffix. Additionally, it makes the opcodes nicer to
write. Finally, and most importantly, it allows us to print the opcode
names in log statements instead of their numbers.

For that reason, we now have the log flag LibServReqs, which logs all
requests to servers in a readable way. As before, LibServ logs all the
kernel-server interaction, but now also in a more readable way.