Avoid try_from at performance-critical places.
Apparently, the compiler is not able to properly optimize the match over
the integer values in try_from, if we perform another match on the
resulting enum variants afterwards. Therefore we instead do a match
over the integer values ourself in the form "o if o == OpCode.into
() => ..." and only use try_from in error cases to retrieve the name of
the operation.