librustm3: improved session removal and creator handling.
Since the creator of a session is only needed by the request handler
when removing a session upon a client request, I've now merged both
functions into one. Therefore, a RequestSession has a method is_dead
that returns an optional creator. By default, it returns None, so that
servers that don't need the session-removal feature can just ignore it.
If a session should be removed, it should return a Some with the
creator id.