librustm3: move all protocol-specifics from Server to RequestHandler.
Previously, Server was still dealing with protocol-specifics, because it
had a default implementation for Handler::exchange and thereby also
required the opcode type. I have now moved all of that over to the
RequestHandler, so that we have a clean separation between the two:
Server is responsible for the communication with the kernel and refers
to a Handler for all details and thus doesn't know anything about
protocols. And RequestHandler implements one Handler that supports
requests from clients and uses handler functions to handle them.
Therefore, RequestHandler also needs to know the supported opcodes and
makes assumptions about the protocol (opcode comes first).