librustm3: let Pipe take ownership of the MemGate.
The MemGate needs to stay alive during the entire pipe communication and
therefore we shouldn't just use it to construct the pipe, but should
use Rust's features to make the requirement clear at the API level.

I think it's fine here to take ownership of the MemGate. We could also
store a reference to let the creator use the MemGate for other things
afterwards. But this makes the API more complicated and I think there
is no real use case for that anyway. And even if we wanted to use the
MemGate for something else afterwards, we could simply derive from it
to give the pipe an exclusive MemGate object.