Start multiplexers lazily from kernel or resource manager.
Previously, we started TileMux on every tile at boot (in parallel to the
kernel) and kept them running until the very end. As we want to be able
to use Linux as the multiplexer, we need to make that more dynamic.

Therefore, the "bootloader" (fpga.py or gem5) now only starts the kernel
on Tile 0. The cores in all other tiles stay turned off. The kernel
then loads TileMux into memory, configures PMP EP0 for that memory
area, and turns on the core. Afterwards, root is started on that tile
as before, which is responsible for all other tiles. That is, before
root starts an activity anywhere, it first initializes the tile in the
same way as the kernel did for root's tile. When allocating a tile from
the resource manager, one can choose whether it should be initialized
(TileMux + PMP EPs) or not. Furthermore, when tiles are freed, they are
now put back into the turned off state (including cache invalidation).

To make that possible, the kernel offers two additional system calls:
- tile_reset: resets a tile and either starts or stops it. The syscall
  requires a root tile cap and a memory cap. If the memory cap is
  valid (and the tile is turned off), it will install this memory cap
  as the first PMP EP and start the core. If the memory cap is invalid
  (and the tile is turned on), it will send TileMux the shutdown
  sidecall, stop the core, reset the tile(including caches), and
  uninstall the PMP EPs.
- tile_mem: gives the caller a memory cap for the tile's internal
  memory, if there is one.

These two calls are leveraged by the resource manager to first load
TileMux into memory (either using tile_mem if it's a tile with internal
memory or by allocating memory elsewhere) and then start TileMux. When
it's done with the tile, it simply resets it by passing an invalid cap
to reset_tile.