Improved address space layout.
Previously, the environment started not on a page boundary on RISC-V,
which caused some trouble here and there. There is actually no good
reason to do that, so that I've changed that now.

I've also moved the receive buffers for TileMux from the end of the
address space to behind the environment. This gives it a fixed location
and simplifies things in the future when we allow different memory
sizes for TileMux.