Allow to specify the multiplexer and enable Linux support again.
We now have multiple new attributes for the dom element in boot scripts
to select the multiplexer for a specific domain, its available memory,
and an initrd. This allows us to pick Linux (or rather, bbl) as the
multiplexer on a specific tile.

The initrd is only required for Linux and is always copied to the end of
the memory area. Note that the DTS is dynamically adjusted for the
current initrd size.

For that reason, I've now removed all the special handling of Linux,
both on hw and on gem5. On the former, the fpga.py now has no longer
support for Linux except that we're using the serial line in this case.
On the latter, we're always using the HiFive platform for RISC-V.
Additionally, we are no longer generating the DTS for gem5, but
maintain it ourself. The platform is basically fixed in our case
anyway, so that we don't gain that much by generating it. And not
generating it at boot allows us to integrate it into bbl so that we
don't have to pass yet another thing to bbl. Or in other words, it works
now on gem5 like it worked on hw before.