m3lx: perform prints via TCU.
Previously, we used the SiFive UART for prints, which has the
disadvantage that prints of M³ (via TCU) and bbl/Linux are partially
out of order. Therefore, we always use the TCU for printing on both
systems now. The implementation is currently in bbl and Linux simply
traps into bbl for printing.

However, for simplicity, input is still done via serial line. The
reasoning is that input is much more complicated, because it needs
interrupts to get notified about new input (for output, we can just
busy wait until it's finished). In theory, we could also use the
TCU-approach for it, but I don't think it's worth the additional
complexity, because input via serial line works fine even when output
is done via TCU.