m3lx: use the same Linux binaries for all targets.
Previously, bbl was target dependent, because of the different DTS files
for gem5 and hw. This is annoying however, because it's pretty easy to
forget that you didn't regenerate bbl after switching targets.

For that reason, I've now decided to pass the DTB at runtime to bbl
instead of integrating it into bbl. This requires a simple patch to
bbl, because we cannot that easily control the registers before
starting bbl (and we would need to do that to put the location of the
DTB into a1). The patch hardcodes the location of the DTB.

To support that, I've added another parameter to the domain to specify
the DTB file. Additionally, the execute.sh script generates the final
DTS by modifying the memory size and initrd location and generates the
DTB. So, with this change we can now actually change the available
memory for Linux in the boot script :)