rust: use features for conditional compilation instead of target_vendor.
This was always a hack that we used to distinguish between different
targets. However, since we can use features for that now (would that
have worked from the beginning?) and it breaks for M³Linux
applications (where we don't control the target spec), we rather use
them.

This is only required for base and m3, because they are used by M³Linux
applications as well, but we use it for all crates now. More
specifically, net and hashmux use it as well and specify it via the
features argument to m3_rust_exe.