librustbase: no write-syscall print on hw.
On hw, the TCU print already ends up in the log.txt file, whereas on
gem5, the TCU print only ends up in the gem5.log. We therefore want to
additionally use the write-syscall print on gem5 to see it both in the
gem5.log and log.txt and only the TCU print on hw.