rustunittests: fixed loop problem in testcliexit test.
The O3 CPU model in gem5 has a bug (at least with x86) in the handling
of jumps to the same instruction. For some reason, it continues with
the next instruction anyway.

To fix that, we need some other instruction in the loop. Since we don't
have to spin here, the best solution is probably a sleep, which is more
desirable than a busy loop anyway.