b: removed mkrootfs command.
Now that we generate the initrd directly before each run, there is no
need to call mkrootfs anymore as it simply runs make in
cross/buildroot, which was already done during the initial
cross-compiler build. If someone wants to add packages to buildroot or
similar, it can do that within the cross directory, as before.