kernel: implement the auto-close feature of sessions.
This should have been done a long time ago, but it wasn't trivial,
because we needed a fully working asynchronous revoke. So far, we
worked around that by letting clients explicitly close sessions that
had been created from existing sessions and therefore without involving
the resource manager. However, this of course doesn't work if the
client is evil or crashes.

Now that the async problems with revoke are gone, we can implement the
auto-close feature relatively easily. The simple idea is to let session
creators decide whether they want to receive a close message if the
session is revoked. If so, the kernel will send that message to the
server on revoke. Note that this is done for any (derived) session
capability so that session sharing is not recommended as any of the
sharers can remove the session for everyone. But we do not share
sessions between clients with any server and I also don't think that
this is a good idea in general.