rust: introduced type VirtAddr.
The idea is to make this a struct in order to avoid that we accidentally
pass another kind of integer for a virtual address. VirtAddr implements
num_traits::PrimInt and therefore can be used like a primitive type,
though.

We also provide explicit conversion functions to convert between
pointers/usize and VirtAddr or goff and VirtAddr. Note that we use u64
as the inner type of VirtAddr, because in a heterogeneous system we
might be talking about virtual addresses on other tiles that have a
different pointer size as our tile (and we currently assume that a u64
should be enough for everyone ;)).