m3lx: support additional receive buffers.
Previously, we only supported the standard receive buffer, because this
was kernel memory provided by the TCU kernel module. Additional receive
buffers are mmap'ed and therefore need additional support.

I've now added support for the MAP sidecall on the Linux side so that
the application can first to the create_map syscall and thereby inform
Linux of the desired mapping. Linux remembers that (physical address,
size, and permissions) and offers a new mmap-type with the TCU device
that maps the last announced mapping to some place within the
applications address space.

Note that the two steps are required, because sidecalls are executed in
interrupt context and therefore cannot block (e.g., required for the
mapping).