Replaced foreign-attribute with mux_info syscall and sidecall.
When we want to start programs on arbitrary tiles, we need to know
whether the tile runs TileMux and therefore supports program loading.
If the tile runs Linux, we cannot load the program but instead only
pass the arguments via environment and use fork+exec on Linux.

For that reason, I've now added a syscall that allows to get information
about the multiplexer on a given tile. The kernel uses a newly
introduced sidecall to ask the multiplexer on that tile for this
information. Currently, it's only the type of multiplexer
(TileMux, Linux, None), but this can be extended in the future.