m3lx: added support for ChildActivity::{run,exec}.
Since we are now able to determine which type of multiplexer is running
on a tile, we can use that to support ChildActivity::{run,exec} on
Linux as well. This works by simply not loading the program, but only
the environment and leave the loading to Linux. For that purpose, we
run the starter application on Linux, which is waiting for new
activities to execute. Whenever that's necessary, it forks itself and
exec's the desired program.

Note that ChildActivity::run only works without ASLR, which is why this
is disabled now. Since this is purely a testing feature, I think it's
fine to disable it.