Add CFIDev tile ISA for tiles with CFI flash
Add a dummy ISA for tiles with CFI flash controller as very simple form
of non-volatile storage. This is mainly intended for gem5 where a CFI
flash emulation is available. In the hardware prototype, a different
kind of flash interface might be easier to integrate.

The flash can be easily read just like regular memory. For writing,
a driver would need to write special commands into the MMIO region.
This is not supported yet.