Add M³RoT implementation
M³RoT is a Root of Trust for M³ based on the Trusted Computing Group
Device Identifier Composition Engine (DICE) specification. Currently it
implements pure Measured Boot, by hashing loaded software and deriving
secrets (Compound Device Identifiers, CDI) based on the software hash
and a root Unique Device Secret (UDS). There is no restriction which
software is allowed to run, but changing the software also results in
different derived secrets.

The RoT runs in a separated RoT tile with a simple CPU core and on-chip
memory (SRAM). Three separated boot layers run sequentially to implement
the RoT boot process:

 1. BROM (Boot ROM): Stored in read-only memory. Loads the first
    measured boot layer from the flash and derives the initial CDI from
    the UDS.

 2. BLAU (Boot Loader Attestation Unit): Long-term assymmetric
    attestation layer. Provides an Ed25519 public key that can be used
    to perform attestation of the loaded software state.

 3. ROSA (Robust Operating System Attestation): M³-specific bootloader.
    Implements the existing M³ boot process that was handled in gem5 so
    far (or the controlling host for the hardware prototype). Loads and
    hashes all boot modules and generates a signed rot-certificate.json
    that describes the started M³ configuration.

These layers are implemented in Rust in a separate Cargo workspace that
is built standalone, exclusively using the Rust toolchain. Unlike the
rest of M³ the C library (musl) is not used. The boot layers are minimal
and can work with 256 KiB of SRAM. The absolute minimum is 128 KiB with
some additional optimizations.

On the M³ side (after the kernel was started) the RoT Service (RoTS)
allows client to derive secrets and derive signatures. The Remote
Attestation Server (RASer) is an example client that implements a simple
challenge-response remote attestation protocol via a text-based TCP
network protocol (or HTTP/0.9).

The common functionality for layers, RoTS and clients is shared in a
separate Root of Trust (rot) library.

The main demo configuration is boot/rots-raser.xml which includes
persistent file system and network stack. boot/rots.xml is a more
minimal configuration that just briefly tests interaction with the RoT
service. Note that to run the RASer demo a network interface must be set
up using the "sudo ip ..." commands printed in the console and a hard
disk specified with the M3_GEM5_HDD environment variable (can use e.g.
build/gem5-riscv-release/default.img).