crypto: kecacc: Add pure Rust backend using sha3 crate
This avoids having to link extra C/C++ code when the hardware
accelerator is missing. It also compiles to much smaller code.

Note that this backend is not fully equivalent to the hardware one
or the emulated one using kecacc-xkcp. In particular there are some
additional validation checks enforced by the sha3 crate, and the state
size is different. For HashMux there is no change in usage, though.