librustbase: Improve safety of C string conversion
When loading the boot modules from the flash we can't be sure that they
were stored there without manipulation, e.g. with the name missing a
null terminator. This means the strlen() call in cstr_to_str() might
access memory out of bounds.

It's probably not critical since it's only a read access but this could
potentially expose secret data in adjacent memory.

For the module/service names, there is no need to trust the loader to
place a valid C string. They are stored in arrays with limited length so
there is an upper bound that should be considered when parsing the C
string.

Switch the custom implementation to use the functions from
core::ffi::CStr, which have additional validation for the conversion.
For now just call .unwrap() to panic rather than causing out of bounds
reads.

TODO: Actually the cstr_to_str() function signature should be changed to
take a core::ffi::c_char as pointer, since i8 is not correct for e.g.
RISC-V where "char" actually refers to an unsigned byte (u8).