m3lx: let Tile::get fail for clone and compat.
The required semantics for "clone" and "compat" are difficult to support
on m3lx. At first, being a clone/compatible tile requires to have the
same multiplexer type, i.e., Linux again. And the semantics of
Tile::get("clone") are that we get a new tile for ourself, which would
require us to boot up a new Linux instance. This takes simply too long
to do that dynamically, IMO. Therefore, the most sensible way to
handle "clone"/"compat" on m3lx is to let it always fail. Meaning,
applications should provide "own" as a fallback.