rust: distinguish non-activated and activated gates via type system.
As it turned out, managing two internal states(non-activated, activated)
within gates leads to quite some performance overhead as we always have
to take into account that the gate might not have been activated yet.
To optimize for the common case at runtime (activated), I've now
decided to distinguish these two states via the type system.

The idea is to have, e.g., a SendCap that acts as a precursor for a
SendGate later. Thus, SendCap only holds the capability and can be
turned into a SendGate via activation. The SendGate is therefore always
activated. The same is true for the RecvGate.

This decision caused some trouble here and there. At first, we took
advantage of the lazy activation during initialization in the sense
that we couldn't have activated a SendGate/RecvGate during
initialization even if we had tried. Specifically, it occurred during
the initialization of OwnActivity. Thus, I've separated the selector
and EP management from OwnActivity (see previous commits) to allow
activation of gates during its initialization.

Another case where we need lazy activation is when gates are obtained
from someone else. As the capability does not exist until the
capability exchanged has been finished by the kernel, we cannot
activate it right away. I've therefore added the type LazyGate that
supports such lazy activation (with a performance cost).