kernel: fixed problem with our SPM emulation on hw.
On the hw platform we emulate SPM by configuring PMP EP0 to some DRAM
region. However, as it turns out we cannot write to the tile's physical
memory (indirectly referring to this DRAM region) until the core in
this tile is running. Unfortunately, we also cannot turn on the core
until we have properly initialized the memory. Thus, we need to write
it to the DRAM region instead.