librustbase: Add support for passing byte slices in IPC
The implementation is very similar to the one that already exists for
strings. The length is prepended, followed by the actual bytes and
padding to make it 64-bit aligned. In fact strings are kind of a special
case of the byte slice encoding (with only valid UTF-8 characters and
additional null termination). Nevertheless, since the string-related IPC
code mentions performance optimizations (e.g. in copy_str_from()) it's
left as-is for now. It could probably be implemented on top of the byte
slice functions.

Unfortunately, using byte slices is somewhat inconvenient in Serde.
While deserialization works correctly, for serialization it typically
prefers to encode an u8 slice as a sequence of u8 bytes, which ends up
as separate u64 words in the current IPC implementation. The only way to
bypass this is using custom Serialize/Deserialize implementations. The
official serde_bytes crate provides functions and wrapper types to avoid
this.

For custom structs this can be avoided by marking the fields
using #serde(with = "serde_bytes")]. When using is.pop()/os.push()
one needs to wrap the &[u8] using serde_bytes::Bytes and the Vec<u8>
using serde_bytes::ByteBuf. Otherwise serde will serialize the bytes
as inefficient sequence.