librustbase: Avoid packed structs where possible
Some of the structs in the base library are marked as repr(packed)
because they are read/written directly to binary representations in
memory or in files, such as the structs representing the ELF header
or the boot environment (boot::Info). repr(packed) ensures there will
be no padding inserted between the fields by the compiler, which may
vary from architecture to architecture.

Unfortunately, repr(packed) also comes with with an suboptimal side
effect: The compiler basically gives up doing alignment entirely.
The struct will effectively have byte-level alignment (1) and the
compiler needs to access all fields byte-wise.

This ends up in suboptimal code at runtime, which is unnecessary if all
of the fields are actually properly aligned.

For example, consider the following Rust code:

    [repr(C, packed)]
    struct Packed {
        num: u64,
    }

    [no_mangle]
    fn read_num(p: &Packed) -> u64 {
        return p.num;
    }

On riscv64 read_num() compiles to roughly the following:

read_num:
        lbu     a1, 1(a0)
        lbu     a2, 0(a0)
        lbu     a3, 2(a0)
        lbu     a4, 3(a0)
        slli    a1, a1, 8
        or      a1, a1, a2
        slli    a3, a3, 16
        slli    a4, a4, 24
        or      a3, a3, a4
        or      a1, a1, a3
        lbu     a2, 5(a0)
        lbu     a3, 4(a0)
        lbu     a4, 6(a0)
        lbu     a0, 7(a0)
        slli    a2, a2, 8
        or      a2, a2, a3
        slli    a4, a4, 16
        slli    a0, a0, 24
        or      a0, a0, a4
        or      a0, a0, a2
        slli    a0, a0, 32
        or      a0, a0, a1
        ret

All 8 bytes must be read one by one and then shifted to the proper
64-bit number.

Without repr(packed) there is just a single instruction:

read_num:
        ld      a0, 0(a0)
        ret

repr(packed) is required when potentially unaligned pointers are
directly accessed via the struct. However, the typical usage in M3
is using something like

  let elf: ElfHeader = TCU::read_obj(...)

where we can guarantee the suitable alignment on the stack. Doing
byte-level accesses in this cases wastes both code size and run time.

Drop the repr(packed) where possible and instead add a const assertion
that ensures the struct has the correct size (meaning that no padding
was inserted). This isn't perfectly reliable but should catch the
problem if some architecture has weird alignment requirements.

There are two remaining uses of repr(packed) in the base library that
are left as-is for now:

 - boot::Tile: Has a TileId (u16) and TileDesc (u64) field which is not
   properly aligned so dropping repr(packed) would change the size.

 - tcu::Header: I wasn't sure if this struct is accessed via potentially
   unaligned pointers.

The last RoT stage uses the ELF structs and boot::{Info, Mem, Mod}.
This commit mainly improves code size: 137 KiB -> 134 KiB (-3 KiB).
There seems to be a very slight improvement in performance as well.