resmng: obtain tile/kmem caps from foreign childs.
This way we can report this within info calls and in general have this
information about foreign childs. As before we are currently trusting
the child to not lie about these selectors (activity, tile, kmem). In
the future we might want to move the activity-creation syscall into the
resource manager and forbid that for normal applications to be sure
about these caps.