build: Build minicov only in coverage build mode
At the moment minicov is always included as dependency of the Rust base
library on RISC-V, even for the "release" build target that does not
have the -Cinstrument-coverage compiler parameter set. Unfortunately,
minicov tends to be a bit "picky" about the build environment, since it
needs e.g. the TARGET_CFLAGS to be set and doesn't work out of the box
on more special RISC-V toolchains such as riscv64imac-unknown-none-elf.
I've also seen it cause trouble for project import in IDEs.

To reduce the chance that one runs into weird minicov errors during
development, make minicov an optional dependency that is only enabled
for the actual "coverage" build. This also allows omitting the
TARGET_CFLAGS handling for the normal "release" build.