CI: Add GitHub Actions workflow for basic compile checks
For now this just runs tries to build M3 in most of the supported
configurations and also runs checks using ./b checkboot, ./b clippy
and ./b fmt. M3Linux is not built at the moment (it already consumes
quite a bit of build time as-is).

The cross compiler is built on the first run and then cached to speed up
subsequent builds. Note however that GitHub deletes the cache again when
no build has run for 7 days. It might be more efficient to download
a pre-built version of the cross compiler.