librustbase: Ensure gem5_writefile sees valid "stdout\0" string
At the moment printing something to the console in gem5 can misbehave in
rare cases. The effect is that the writes don't end up being displayed
at all or only partially and instead weird garbage files are created in
the run/ directory, e.g.

  "run/stdod 'app' exited with exitcode InvArgs"

It looks like a part of the file target string for gem5_writefile
("stdout\0") is not read correctly. This seems to happen because parts
of the write operation end up in the store buffer of the CPU (or
something similar). gem5 reads these strings by sending out functional
requests on the CPU data port, which will bypass the internal store
buffer of the CPU.

Prevent this from happening by forcing the CPU to write out the string
fully (using a fence) before the gem5 pseudo instruction is executed.

Note that placing the fence immediately before the gem5_writefile
wouldn't solve the problem because the fence will typically only affect
ordering of preceding memory accesses with regards to later memory
accesses. The execution of the gem5 pseudo instruction is not a memory
access. However, it works fine when placed before the read_volatile().