librustbase: Avoid copying "stdout\0" to the stack for gem5_writefile
Touching the string with the read_volatile() accesses should be
sufficient to make sure it's mapped, even if the string is located in
.rodata (or similar) instead of the stack. The C++ version in Machine.cc
doesn't copy the string to the stack either.

In general, reusing the string constant also avoids the problem with the
store buffer so the atomic::fence() could be removed again. However,
strictly speaking even the buffer given as function parameter could be
partially buffered in the CPU. This means there is still a tiny little
chance that the fence is needed.