librustm3: Avoid use of core_intrinsics
core::mem::transmute() is a re-export of core::intrinsics::transmute()
that works on stable Rust as well.

TODO: Should likely make "base" and other parts of M3 use the stable
version as well, but they currently still use other intrinsics that do
not have a stable counterpart.

This fixes the following warning on Rust 2023-12-21:

warning: the feature `core_intrinsics` is internal to the compiler or standard library
  --> libs/rust/m3impl/src/lib.rs:19:12
   |
19 | #![feature(core_intrinsics)]
   |            ^^^^^^^^^^^^^^^
   |
   = note: using it is strongly discouraged
   = note: `#[warn(internal_features)]` on by default

warning: `m3impl` (lib) generated 1 warning