rot: Add minimal μBROM implemented in RISC-V assembly
μBROM is a minimal re-implementation of BROM purely in RISC-V assembly.
Since BROM would be immutable in real hardware this results in a
significantly smaller Trusted Computing Base (TCB). The complex compiler
toolchain does not have to be trusted to compile code correctly.

Since the immutable part of M³RoT is minimal by design μBROM also
consists only of roughly 100 instructions (mostly just triggering
commands in the TCU and the hash accelerator). The handling of secrets
is clearly marked and limited to a few instructions.

It is currently a bit harder to maintain than the Rust version of BROM
because some MMIO offsets to the TCU and the hash accelerator are
hardcoded. The effort for updates could be reduced slightly by having
shared definitions in a C header.

To test μBROM, replace "brom" with "ubrom" in the <rot> element of the
boot.xml.