Support in-memory endpoints.
The TCU now provides registers to specify an external memory region for
endpoints. The region can be arbitrarily large, leading to an unbounded
number of endpoints. However, TCUs can still have internal endpoints in
which case the EPS_SIZE register holds the size of the internal EP
region in bytes and EPS_ADDR is ignored.

For security reasons, the kernel uses a dedicated memory area for all EP
memory regions and keeps this memory area to itself. Userspace specifies
the number of EPs for a given tile with the tile_reset syscall. The
kernel allocates a correspondingly large region within the EP memory
area and clears it before usage. The number of available EPs (for tiles
with external and internal EPs) can be obtained with the tile_info
syscall.

The resource manager therefore takes advantage of tile_reset to specify
the number of EPs according to the EPs required for the applications
that are supposed to run on this tile.