kernel: ensure that we don't send further msgs to shutdown TileMux's.
This problem occurs if we first send the shutdown sidecall to TileMux
and then switch to another thread until the response is available. Since
we only mark TileMux as "not available" after the response has been
received, other threads might send further messages to this TileMux
instance in the meantime.

TileMux cannot and should not receive further messages as it completely
shuts itself off after sending the response. We therefore now maintain
another flag in the tilemux module that keeps track of whether we
already sent the shutdown sidecall. To prevent that we then fail the "is
TileMux initialized" call for the shutdown sidecall, we disable the
check for shutdown.