Renamed BOOM/ROCKET tile attributes to PERF/EFFI.
The rational is that these names should be more generic as we might have
different cores on every platform and want to keep configuration files
as generic as possible. And in the end, we primary care about the
performance/efficiency properties of these cores and not on it being a
BOOM/ROCKET core specifically.