rot: Implement basic caching extension as performance improvement
Caching is an extension to M³RoT that improves startup performance
significantly when the same software state is booted again. When the
public key and signature cached in the flash corresponds to the loaded
software state, computation of asymmetric cryptography is skipped
completely. When the installed software is changed, new public keys and
signatures are re-computed and should be stored for the next startup.

Three symmetric Message Authentication Codes (MACs) are used to check
that the cached keys and signatures conform to the actually loaded
software state. They are loaded from the RoT configuration.

Right now the cache fields still need to be hardcoded in gem5 because
there is no driver for persistently writing to the flash in M³ yet. The
boot layers currently just print the new MAC on mismatch, which should
then be set for the RoT configuration in gem5.