resmng: map credits=0 to unlimited credits.
While we need a different way to represent unlimited credits in the TCU
(as 0 is a valid current number of credits), giving a SendGate 0 credits
in a boot script makes little sense as it would not be able to gain more
credits.

To avoid that users need to know the exact value of the UNLIM_CREDITS
constant, one can now specify credits="0" to give unlimited credits.